+++ builtin/__init__.py
@@ -1,0 +1,2 @@
+import miniserve.builtin.GET
+import miniserve.builtin.POST
